backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.44
asyncmy==0.2.9
aiosqlite==0.21.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
bcrypt>=3.2.0,<4.0.0
python-jose==3.3.0
python-dotenv==1.0.1
pydantic==2.12.2
pydantic-settings==2.11.0
python-multipart==0.0.9
openai==2.3.0
httpx==0.28.1
email-validator==2.1.1
cryptography>=41.0.0
redis==5.0.7
libsql-client==0.3.1
ollama==0.6.0
langchain-text-splitters==0.3.11
greenlet==3.2.4
python-docx==1.1.2
